[TASK-3] Reorganize project structure: rename packages and add data directory
