+++ bolatbyek.nazgul/task-3/cmd/service/main.go
@@ -4,10 +4,10 @@ import (
 	"flag"
 	"fmt"
 
-	"github.com/Nazkaaa/task-3/internal/cbr"
 	"github.com/Nazkaaa/task-3/internal/config"
-	"github.com/Nazkaaa/task-3/internal/convert"
-	"github.com/Nazkaaa/task-3/internal/output"
+	"github.com/Nazkaaa/task-3/internal/jsonwriter"
+	"github.com/Nazkaaa/task-3/internal/models"
+	"github.com/Nazkaaa/task-3/internal/xmlparser"
 )
 
 func run(configPath string) error {
@@ -21,19 +21,19 @@ func run(configPath string) error {
 		return fmt.Errorf("error creating output directory: %w", err)
 	}
 
-	valCurs, err := cbr.ParseXML(cfg.InputFile)
+	valCurs, err := xmlparser.ParseXML(cfg.InputFile)
 	if err != nil {
 		return fmt.Errorf("error parsing XML: %w", err)
 	}
 
-	currencies := convert.ConvertAndSort(valCurs)
+	currencies := models.ConvertAndSort(valCurs)
 
 	var outputCurrencies []interface{}
 	for _, c := range currencies {
 		outputCurrencies = append(outputCurrencies, c)
 	}
 
-	err = output.SaveToJSON(outputCurrencies, cfg.OutputFile)
+	err = jsonwriter.SaveToJSON(outputCurrencies, cfg.OutputFile)
 	if err != nil {
 		return fmt.Errorf("error saving to JSON: %w", err)
 	}
