+++ bolatbyek.nazgul/task-3/data/currencies.json
@@ -1,0 +1,18 @@
+[
+  {
+    "num_code": 840,
+    "char_code": "USD",
+    "value": 30.9436
+  },
+  {
+    "num_code": 826,
+    "char_code": "GBP",
+    "value": 43.8254
+  },
+  {
+    "num_code": 36,
+    "char_code": "AUD",
+    "value": 16.0102
+  }
+]
+
