+++ bolatbyek.nazgul/task-3/internal/jsonwriter/jsonwriter.go
@@ -1,4 +1,4 @@
-package output
+package jsonwriter
 
 import (
 	"encoding/json"
