+++ bolatbyek.nazgul/task-3/internal/xmlparser/xmlparser.go
@@ -1,4 +1,4 @@
-package cbr
+package xmlparser
 
 import (
 	"bytes"
@@ -6,22 +6,11 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/Nazkaaa/task-3/internal/models"
 	"golang.org/x/net/html/charset"
 )
 
-type ValCurs struct {
-	XMLName xml.Name `xml:"ValCurs"`
-	Valutes []Valute `xml:"Valute"`
-}
-
-type Valute struct {
-	XMLName  xml.Name `xml:"Valute"`
-	NumCode  string   `xml:"NumCode"`
-	CharCode string   `xml:"CharCode"`
-	Value    string   `xml:"Value"`
-}
-
-func ParseXML(filePath string) (*ValCurs, error) {
+func ParseXML(filePath string) (*models.ValCurs, error) {
 	xmlData, err := os.ReadFile(filePath)
 	if err != nil {
 		return nil, fmt.Errorf("failed to read XML file %q: %w", filePath, err)
@@ -30,7 +19,7 @@ func ParseXML(filePath string) (*ValCurs, error) {
 	decoder := xml.NewDecoder(bytes.NewReader(xmlData))
 	decoder.CharsetReader = charset.NewReaderLabel
 
-	var valCurs ValCurs
+	var valCurs models.ValCurs
 	if err := decoder.Decode(&valCurs); err != nil {
 		return nil, fmt.Errorf("failed to decode XML from file %q: %w", filePath, err)
 	}
